Increase window capacity to uint16
- Defaults to 1024, instead of 10
- Gives us a ~50x increase in speed

+++ flu/messages/parse.go
@@ -60,7 +60,7 @@ func Parse(data []byte) (msg Message, err error) {
 	case openLineRequest:
 		sha1Hash := reader.readSha1Hash()
 		chunk := reader.readUint16()
-		cap := reader.readByte()
+		cap := reader.readUint16()
 		return &OpenConnectionRequest{
 			Sha1Hash:  sha1Hash,
 			Chunk:     chunk,
